fix(Avalonia): do not use SkiaSharp NoDependencies variant
Fixes #212

Hey, we're pulling back on using `SkiaSharp.NativeAssets.Linux.NoDependencies`. Turns out, it just wasn't playing nice with the **Inter font** that we've got baked into the app. And honestly, if it can't handle a common font like Inter, it's pretty much useless to us for what we need.

By switching back to the regular SkiaSharp setup, we're not only fixing the Inter font issue now, but we're also setting things up so that, down the road, you'll be able to change the font SnapX uses yourself.