build(GitVersion): do not increment solely by branch
Before this change, it was always a Minor increment because of the fact the change was on develop. Now, it's based off of commit messages and tagged commits.

I fully acknowledge the develop branch is still v0.4.0 after this change. It is because of commit 262768bc91850cd5b116287123833a90396a177b , otherwise, it'd be v0.3.1